Refactor modules for season
