+++ spec/league_stats_spec.rb
@@ -1,5 +1,6 @@
 require './lib/stat_tracker'
 require './lib/league_stats'
+require './lib/league_stats_module'
 require 'simplecov'
 SimpleCov.start
 require 'csv'
